LPS-178949 Reorganizing methods from SegmentsDisplayContext to SegmentsManagementToolbarDisplayContext class
